:mag: Upgraded Telegram-invite due platform changes
Telegram has updated the presentation of invitelinks and some of the old
ones may have broken. Even though my taps on them take me to the right
group, I suspect that this is because the outdated links only work for
those who have already joined because of the connection between the link
and the site in the history.

I could also use the name links https://t.me/TulpaWiki or
https://tulpawiki.t.me/, but I want to separate the newcomer channels
with different links.